pyproject.toml
[project]
name = "abliterix"
version = "1.0.1"
description = "Automated model steering and alignment adjustment via LoRA-based optimization"
readme = "README.md"
license = "AGPL-3.0-or-later"
authors = [
    { name = "Wangzhang Wu", email = "wangzhangwu1216@gmail.com" }
]
requires-python = ">=3.10"
keywords = ["llm", "model-steering", "alignment", "lora"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: GPU",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "accelerate~=1.10",
    "bitsandbytes~=0.45",
    "datasets~=4.0",
    "hf-transfer~=0.1",
    "huggingface-hub~=1.6",
    "kernels~=0.11",
    "optuna~=4.5",
    "peft~=0.14",
    "psutil~=7.1",
    "pydantic-settings~=2.10",
    "questionary~=2.1",
    "rich~=14.1",
    "transformers~=5.3",
]

[project.optional-dependencies]
research = [
    "geom-median~=0.1",
    "imageio~=2.37",
    "matplotlib~=3.10",
    "numpy~=2.2",
    "pacmap~=0.8",
    "scikit-learn~=1.7",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "ruff>=0.14.5",
    "ty>=0.0.5",
]

[project.urls]
Homepage = "https://github.com/wuwangzhang1216/abliterix"
Documentation = "https://github.com/wuwangzhang1216/abliterix"
Repository = "https://github.com/wuwangzhang1216/abliterix.git"
Issues = "https://github.com/wuwangzhang1216/abliterix/issues"
Changelog = "https://github.com/wuwangzhang1216/abliterix/releases"

[project.scripts]
abliterix = "abliterix.cli:main"

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "abliterix"

[tool.ruff.lint.per-file-ignores]
"scripts/*.py" = ["E402"]
"tests/*.py" = ["E402"]
